requirements.txt
numpy==1.26.4
scipy==1.14.1
pandas==2.2.2
PyYAML==6.0.2
scikit_learn==1.5.1
dvc==3.59.1
lightgbm==4.6.0
